refactor(server): extract render marshalling into tools/_marshal.py (#183 part B)
render_view and render_drawing each assembled a list[ImageContent|TextContent]
(tempfile fallback + [SEND:] markers + base64) inline — the last real business
logic in the tool wrappers. Move it byte-identically into
tools/_marshal.py::marshal_render_view / marshal_render_drawing; the wrappers
now just call the worker and return the marshalled result.

server.py drops 741 -> 652 lines (891 -> 652 across parts A+B) and is now
registration + the configure() setter. No behavior change — guarded by
test_outcomes.py::test_mcp_render_returns_image_and_file_path (image type, PNG
magic bytes, [SEND:] marker, file-on-disk). Gate: mypy + ruff clean, 600 passed.

Closes #183
